requirements.txt
telethon
fastapi
uvicorn
psycopg2-binary
sqlalchemy
python-dotenv
ultralytics
dbt-postgres
dbt
dagster
pytest